+++ Tests/SentryTests/SentrySwiftTests.swift
@@ -7,7 +7,7 @@ class SentrySwiftTests: XCTestCase {
     override func setUp() {
         super.setUp()
         let fileManager = try! SentryFileManager(dsn: SentryDsn(string: "https://username:password@app.getsentry.com/12345"), andCurrentDateProvider: TestCurrentDateProvider())
-        fileManager.deleteAllStoredEventsAndEnvelopes()
+        fileManager.deleteAllEnvelopes()
         fileManager.deleteAllFolders()
         _ = SentrySDK.start { options in
             options.dsn = "https://username:password@app.getsentry.com/12345"